refactor(pkg): split module to "browser" and "node"
