+++ test/mocha.opts
@@ -1,1 +1,1 @@
---compilers ts:ts-node/register
+--require ts-node/register
